+++ xfrm_policy_linux_test.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 package netlink
 
 import (
